backend/pyproject.toml
[project]
name = "minimal-local-ai-agent-backend"
version = "0.1.0"
description = "Minimal production-minded local AI agent backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.34.0",
  "gunicorn>=23.0.0",
  "sqlalchemy>=2.0.36",
  "greenlet>=3.1.1",
  "aiosqlite>=0.20.0",
  "alembic>=1.14.0",
  "jinja2>=3.1.4",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "python-dotenv>=1.0.1",
  "openai>=1.58.1",
  "httpx>=0.28.1",
  "opentelemetry-api>=1.28.2",
  "opentelemetry-sdk>=1.28.2",
  "opentelemetry-instrumentation-fastapi>=0.49b2",
  "opentelemetry-exporter-otlp>=1.28.2",
  "sse-starlette>=2.1.3",
  "prometheus-client>=0.21.1",
  "slowapi>=0.1.9"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

